refacttor: Update ids in Change Report area to use Id types

+++ plugins/org.eclipse.osee.orcs.db.test/src/org/eclipse/osee/orcs/db/internal/change/ComputeNetChangeTest.java
@@ -39,56 +39,56 @@ public class ComputeNetChangeTest {
       List<TestData> data = new ArrayList<>();
 
       // New Or Introduced
-      data.add(createTest(1, entry(3L, NEW), entry(4L, NEW), entry(5L, MODIFIED), null, entry(5L, NEW), false));
-      data.add(
-         createTest(2, entry(3L, NEW), entry(6L, INTRODUCED), entry(7L, MODIFIED), null, entry(7L, INTRODUCED), false));
-      data.add(createTest(3, entry(3L, NEW), null, entry(7693330L, INTRODUCED), entry(7693330L, NEW),
+      data.add(createTest(1L, entry(3L, NEW), entry(4L, NEW), entry(5L, MODIFIED), null, entry(5L, NEW), false));
+      data.add(createTest(2L, entry(3L, NEW), entry(6L, INTRODUCED), entry(7L, MODIFIED), null, entry(7L, INTRODUCED),
+         false));
+      data.add(createTest(3L, entry(3L, NEW), null, entry(7693330L, INTRODUCED), entry(7693330L, NEW),
          entry(7693330L, INTRODUCED), false));
 
       // Modified Once
-      data.add(createTest(4, entry(10L, MODIFIED), null, entry(11L, MODIFIED), entry(10L, MODIFIED),
+      data.add(createTest(4L, entry(10L, MODIFIED), null, entry(11L, MODIFIED), entry(10L, MODIFIED),
          entry(11L, MODIFIED), false));
 
       // Modified Twice
-      data.add(createTest(5, entry(10L, NEW), entry(11L, MODIFIED), entry(12L, MODIFIED), entry(10L, NEW),
+      data.add(createTest(5L, entry(10L, NEW), entry(11L, MODIFIED), entry(12L, MODIFIED), entry(10L, NEW),
          entry(12L, MODIFIED), false));
 
       // Removal - new/intro and deleted
-      data.add(createTest(6, entry(3L, NEW), entry(1L, NEW), entry(2L, DELETED), null, null, true));
-      data.add(createTest(7, entry(3L, NEW), entry(2L, INTRODUCED), entry(3L, DELETED), null, null, true));
-      data.add(createTest(8, entry(3L, NEW), entry(4L, NEW), entry(5L, ARTIFACT_DELETED), null, null, true));
-      data.add(createTest(9, entry(3L, NEW), entry(6L, INTRODUCED), entry(7L, ARTIFACT_DELETED), null, null, true));
-      data.add(createTest(10, entry(3L, NEW), null, entry(21345L, NEW), entry(21345L, NEW), null, true));
-      data.add(createTest(11, entry(3L, NEW), null, entry(1L, NEW), entry(2L, MODIFIED), null, entry(1L, NEW), false));
-      data.add(createTest(12, entry(3L, NEW), null, entry(1L, INTRODUCED), entry(2L, DELETED), null, null, true));
+      data.add(createTest(6L, entry(3L, NEW), entry(1L, NEW), entry(2L, DELETED), null, null, true));
+      data.add(createTest(7L, entry(3L, NEW), entry(2L, INTRODUCED), entry(3L, DELETED), null, null, true));
+      data.add(createTest(8L, entry(3L, NEW), entry(4L, NEW), entry(5L, ARTIFACT_DELETED), null, null, true));
+      data.add(createTest(9L, entry(3L, NEW), entry(6L, INTRODUCED), entry(7L, ARTIFACT_DELETED), null, null, true));
+      data.add(createTest(10L, entry(3L, NEW), null, entry(21345L, NEW), entry(21345L, NEW), null, true));
+      data.add(createTest(11L, entry(3L, NEW), null, entry(1L, NEW), entry(2L, MODIFIED), null, entry(1L, NEW), false));
+      data.add(createTest(12L, entry(3L, NEW), null, entry(1L, INTRODUCED), entry(2L, DELETED), null, null, true));
 
       // Undelete then delete again (Resurrected) not ignored
-      data.add(createTest(13, entry(2L, DELETED), entry(3L, MODIFIED), entry(4L, DELETED), entry(5L, DELETED),
+      data.add(createTest(13L, entry(2L, DELETED), entry(3L, MODIFIED), entry(4L, DELETED), entry(5L, DELETED),
          entry(4L, DELETED), false));
-      data.add(createTest(14, entry(2L, ARTIFACT_DELETED), null, entry(4L, ARTIFACT_DELETED),
+      data.add(createTest(14L, entry(2L, ARTIFACT_DELETED), null, entry(4L, ARTIFACT_DELETED),
          entry(5L, ARTIFACT_DELETED), entry(4L, ARTIFACT_DELETED), false));
 
       // Delete Cases -
-      data.add(createTest(15, entry(10L, MODIFIED), null, entry(10L, DELETED), entry(10L, MODIFIED),
+      data.add(createTest(15L, entry(10L, MODIFIED), null, entry(10L, DELETED), entry(10L, MODIFIED),
          entry(10L, DELETED), false));
-      data.add(createTest(16, entry(330L, NEW), null, entry(331L, ARTIFACT_DELETED), entry(330L, NEW),
+      data.add(createTest(16L, entry(330L, NEW), null, entry(331L, ARTIFACT_DELETED), entry(330L, NEW),
          entry(331L, ARTIFACT_DELETED), false));
-      data.add(createTest(17, entry(330L, NEW), entry(331L, ARTIFACT_DELETED), entry(331L, ARTIFACT_DELETED),
+      data.add(createTest(17L, entry(330L, NEW), entry(331L, ARTIFACT_DELETED), entry(331L, ARTIFACT_DELETED),
          entry(330L, NEW), entry(331L, ARTIFACT_DELETED), false));
 
       // Parent to Child Intro
-      data.add(createTest(18, null, entry(10L, INTRODUCED), entry(11L, MODIFIED), entry(10L, MODIFIED),
+      data.add(createTest(18L, null, entry(10L, INTRODUCED), entry(11L, MODIFIED), entry(10L, MODIFIED),
          entry(11L, MODIFIED), false));
 
       // Attribute Modified/New
-      data.add(createTest(19, null, null, entry(11L, MODIFIED), null, entry(11L, NEW), false));
+      data.add(createTest(19L, null, null, entry(11L, MODIFIED), null, entry(11L, NEW), false));
 
       // Test Merge Items
-      data.add(createTest(20, null, null, entry(12L, MODIFIED), null, entry(13L, MERGED), entry(13L, MERGED), false));
-      data.add(createTest(21, null, null, entry(12L, NEW), null, entry(14L, MERGED), entry(14L, MERGED), false));
-      data.add(createTest(22, entry(96915L, MODIFIED), entry(7290448L, MODIFIED), entry(7865315L, DELETED),
+      data.add(createTest(20L, null, null, entry(12L, MODIFIED), null, entry(13L, MERGED), entry(13L, MERGED), false));
+      data.add(createTest(21L, null, null, entry(12L, NEW), null, entry(14L, MERGED), entry(14L, MERGED), false));
+      data.add(createTest(22L, entry(96915L, MODIFIED), entry(7290448L, MODIFIED), entry(7865315L, DELETED),
          entry(7432082L, MODIFIED), entry(7865315L, MERGED), entry(7865315L, DELETED), false));
-      data.add(createTest(21, entry(96915L, MODIFIED), entry(7290448L, MODIFIED), entry(7865315L, ARTIFACT_DELETED),
+      data.add(createTest(21L, entry(96915L, MODIFIED), entry(7290448L, MODIFIED), entry(7865315L, ARTIFACT_DELETED),
          entry(7432082L, MODIFIED), entry(7865315L, MERGED), entry(7865315L, ARTIFACT_DELETED), false));
 
       return data;
@@ -127,18 +127,18 @@ public class ComputeNetChangeTest {
       List<ChangeItem> items = new ArrayList<>();
 
       // Source to Non-Parent commit
-      items.add(ChangeTestUtility.createItem(3, entry(10L, MODIFIED), null, entry(11L, MODIFIED), null, null));
+      items.add(ChangeTestUtility.createItem(3L, entry(10L, MODIFIED), null, entry(11L, MODIFIED), null, null));
 
       Callable<List<ChangeItem>> callable = new ComputeNetChangeCallable(items);
       callable.call();
    }
 
-   private static TestData createTest(int itemId, ChangeVersion base, ChangeVersion first, ChangeVersion current, ChangeVersion destination, ChangeVersion expected, boolean isRemoved) {
+   private static TestData createTest(Long itemId, ChangeVersion base, ChangeVersion first, ChangeVersion current, ChangeVersion destination, ChangeVersion expected, boolean isRemoved) {
       return new TestData(ChangeTestUtility.createItem(itemId, base, first, current, destination, null), expected,
          isRemoved);
    }
 
-   private static TestData createTest(int itemId, ChangeVersion base, ChangeVersion first, ChangeVersion current, ChangeVersion destination, ChangeVersion net, ChangeVersion expected, boolean isRemoved) {
+   private static TestData createTest(Long itemId, ChangeVersion base, ChangeVersion first, ChangeVersion current, ChangeVersion destination, ChangeVersion net, ChangeVersion expected, boolean isRemoved) {
       return new TestData(ChangeTestUtility.createItem(itemId, base, first, current, destination, net), expected,
          isRemoved);
    }
